+++ bindings/python/examples/client/06_simple_block.py
@@ -26,7 +26,7 @@ unsigned_block = client.build_basic_block(
     TaggedDataPayload(
         utf8_to_hex("tag"),
         utf8_to_hex("data")))[0]
-signed_block = secret_manager.sign_block(unsigned_block, chain)
-block_id = client.post_block(signed_block)
+block = secret_manager.sign_block(unsigned_block, chain)
+block_id = client.post_block(block)
 
 print(f'Empty block sent: {os.environ["EXPLORER_URL"]}/block/{block_id}')
