+++ bindings/python/iota_sdk/client/_high_level_api.py
@@ -4,7 +4,7 @@
 from typing import List, Optional
 from dataclasses import dataclass
 from abc import ABCMeta, abstractmethod
-from iota_sdk.types.block.signed_block import Block
+from iota_sdk.types.block.block import Block
 from iota_sdk.types.common import CoinType, HexStr, json
 from iota_sdk.types.output_metadata import OutputWithMetadata
 from iota_sdk.types.output_id import OutputId
