+++ bindings/python/iota_sdk/types/block/body/basic.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import List, Optional
 from dataclasses_json import config
-from iota_sdk.types.block.block import BlockBodyType
+from iota_sdk.types.block.body.kind import BlockBodyType
 from iota_sdk.types.common import HexStr, json
 from iota_sdk.types.payload import Payload
 
@@ -13,7 +13,7 @@ from iota_sdk.types.payload import Payload
 @json
 @dataclass
 class BasicBlockBody:
-    """A `BasicBlockBody` is the most common type of block body used to issue various kinds of payloads such as transactions
+    """A Basic Block Body is the most common type of block body used to issue various kinds of payloads such as transactions
     at the cost of burning Mana.
 
     Attributes:
