+++ bindings/python/iota_sdk/types/block/body/validation.py
@@ -4,7 +4,7 @@
 from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import List, Optional
-from iota_sdk.types.block.block import BlockBodyType
+from iota_sdk.types.block.body.kind import BlockBodyType
 from iota_sdk.types.common import HexStr, json
