+++ bindings/python/iota_sdk/types/block/metadata.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 from typing import Optional
 from iota_sdk.types.common import HexStr, json
 # TODO rename change to Block
-from iota_sdk.types.block.signed_block import Block
+from iota_sdk.types.block.block import Block
 
 
 @json
